fix code style

+++ src/operator/quantization/mkldnn/mkldnn_quantize-inl.h
@@ -1,4 +1,4 @@
- /*******************************************************************************
+/*
  * Copyright 2016-2017 Intel Corporation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,7 +18,7 @@
  * \file mkldnn_quantize-inl.h
  * \brief
  * \author Wenting Jiang
-*/
+ */
 #ifndef MXNET_OPERATOR_QUANTIZATION_MKLDNN_MKLDNN_QUANTIZE_INL_H_
 #define MXNET_OPERATOR_QUANTIZATION_MKLDNN_MKLDNN_QUANTIZE_INL_H_
 #include <string>
@@ -32,16 +32,15 @@ namespace op {
 
 template<typename SrcType, typename DstType>
 void MKLDNNQuantizeComputeKer(const std::vector<TBlob>& inputs,
-                           const std::vector<TBlob>& outputs,
-                           const QuantizeParam& param) {
+                              const std::vector<TBlob>& outputs,
+                              const QuantizeParam& param) {
   using namespace mshadow;
   using namespace mxnet_op;
   using red::limits::MaxValue;
   using red::limits::MinValue;
-
   // check shapes
-  int i_dim = inputs[0].ndim();
-  int o_dim = outputs[0].ndim();
+  size_t i_dim = inputs[0].ndim();
+  size_t o_dim = outputs[0].ndim();
   CHECK_EQ(i_dim, o_dim);
   int total_len = 1;
   memory::dims tensor_shape;
@@ -50,7 +49,6 @@ void MKLDNNQuantizeComputeKer(const std::vector<TBlob>& inputs,
     total_len *= inputs[0].size(i);
   }
   tensor_shape.push_back(total_len);
- 
   float real_range = MaxAbs(*inputs[1].dptr<float>(), *inputs[2].dptr<float>());
   float quantized_range = MaxAbs(MaxValue<DstType>(), MinValue<DstType>());
   *outputs[1].dptr<float>() = *inputs[1].dptr<float>();
@@ -68,9 +66,7 @@ void MKLDNNQuantizeComputeKer(const std::vector<TBlob>& inputs,
   std::vector<float> scales = {scale};
   attr.set_output_scales(mask, scales);
   attr.set_int_output_round_mode(round_nearest);
- 
   mkldnn::engine cpu_engine = mxnet::CpuEngine::Get()->get_engine();
-  
   auto i_mpd = memory::primitive_desc({tensor_shape,
                                       (mkldnn::memory::data_type)data_type_enum<SrcType>::type,
                                        memory::format::x},
@@ -79,20 +75,18 @@ void MKLDNNQuantizeComputeKer(const std::vector<TBlob>& inputs,
                                       (mkldnn::memory::data_type)data_type_enum<DstType>::type,
                                        memory::format::x},
                                        cpu_engine);
-  
   auto reorder_pd  = reorder::primitive_desc(i_mpd, o_mpd, attr);
   auto input = memory(i_mpd, inputs[0].dptr<SrcType>());
   auto output = memory(o_mpd, outputs[0].dptr<DstType>());
-  
   auto r = reorder(reorder_pd, input, output);
   stream(stream::kind::lazy).submit({r}).wait();
 }
 
 void MKLDNNQuantizeCompute(const nnvm::NodeAttrs& attrs,
-                     const OpContext& ctx,
-                     const std::vector<TBlob>& inputs,
-                     const std::vector<OpReqType>& req,
-                     const std::vector<TBlob>& outputs) {
+                           const OpContext& ctx,
+                           const std::vector<TBlob>& inputs,
+                           const std::vector<OpReqType>& req,
+                           const std::vector<TBlob>& outputs) {
   const QuantizeParam& param = nnvm::get<QuantizeParam>(attrs.parsed);
   if (param.out_type == mshadow::kUint8) {
     MKLDNNQuantizeComputeKer<float, uint8_t>(inputs, outputs, param);
